vetiver_requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=vetiver_requirements.txt /tmp/tmp46ja0_6m.in
#
gcsfs
annotated-doc==0.0.3
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   starlette
appdirs==1.4.4
    # via pins
build==1.3.0
    # via pip-tools
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via
    #   pip-tools
    #   rsconnect-python
    #   uvicorn
databackend==0.0.3
    # via pins
fastapi==0.121.0
    # via vetiver
fsspec==2025.7.0
    # via pins
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via vetiver
humanize==4.14.0
    # via pins
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.7.0
    # via pins
importlib-resources==6.5.2
    # via pins
jinja2==3.1.6
    # via pins
joblib==1.5.2
    # via
    #   pins
    #   scikit-learn
    #   vetiver
markupsafe==3.0.3
    # via jinja2
narwhals==2.10.2
    # via plotly
nest-asyncio==1.6.0
    # via vetiver
numpy==2.3.4
    # via
    #   pandas
    #   scikit-learn
    #   scipy
    #   vetiver
packaging==25.0
    # via
    #   build
    #   plotly
pandas==2.3.3
    # via
    #   pins
    #   vetiver
pins==0.9.1
    # via vetiver
pip-tools==7.5.1
    # via vetiver
plotly==6.4.0
    # via vetiver
pydantic==2.12.4
    # via
    #   fastapi
    #   vetiver
pydantic-core==2.41.5
    # via pydantic
pyjwt==2.10.1
    # via rsconnect-python
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.1
    # via vetiver
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via pins
requests==2.32.5
    # via
    #   pins
    #   vetiver
rsconnect-python==1.28.0
    # via vetiver
scikit-learn==1.7.2
    # via
    #   -r /tmp/tmp46ja0_6m.in
    #   vetiver
scipy==1.16.3
    # via scikit-learn
semver==3.0.4
    # via rsconnect-python
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.49.3
    # via fastapi
threadpoolctl==3.6.0
    # via scikit-learn
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   pins
    #   pydantic
    #   pydantic-core
    #   rsconnect-python
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
uvicorn==0.38.0
    # via vetiver
vetiver==0.2.6
    # via -r /tmp/tmp46ja0_6m.in
wheel==0.45.1
    # via pip-tools
xxhash==3.6.0
    # via pins
zipp==3.23.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
